Fix obvious errors in code.
Error #1: The default latency should be the one to the data center. So if a cache has a higher latency then the data center, then it isn't selected. As per the problem documentation, L(c) < L(d), however it is still best to include this.

Error #2: Lowest latency wasn't being reset. Meaning as per PHP array mapping, the last thrown cache would be the one to get thrown. Hence if a video couldn't be stored in the last one, it wouldn't be stored at all. This is now fixed.